task

+++ PhotosService/Controllers/PhotosApiController.cs
@@ -9,7 +9,13 @@ using PhotosService.Models;
 
 namespace PhotosService.Controllers
 {
+    using System.IdentityModel.Tokens.Jwt;
+    using Microsoft.AspNetCore.Authorization;
+    using Microsoft.AspNetCore.Http.Extensions;
+    using Services;
+
     [ApiController]
+    [Authorize]
     [Route("api/photos")]
     public class PhotosApiController : Controller
     {
@@ -23,9 +29,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetPhotos(string ownerId)
+        public async Task<IActionResult> GetPhotos(string ownerId, [FromHeader] JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != ownerId)
+                return Forbid();
+            
             var photoEntities = (await photosRepository.GetPhotosAsync(ownerId)).ToList();
+
             var photos = mapper.Map<List<PhotoDto>>(photoEntities);
             foreach(var photo in photos)
                 photo.Url = GeneratePhotoUrl(photo);
@@ -33,11 +43,13 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet("{id}/meta")]
-        public async Task<IActionResult> GetPhotoMeta(Guid id)
+        public async Task<IActionResult> GetPhotoMeta(Guid id, [FromHeader] JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
 
             var photo = mapper.Map<PhotoDto>(photoEntity);
             photo.Url = GeneratePhotoUrl(photo);
@@ -45,8 +57,35 @@ namespace PhotosService.Controllers
         }
 
         [HttpGet("{id}/content")]
-        public async Task<IActionResult> GetPhotoContent(Guid id)
+        public async Task<IActionResult> GetPhotoContent(Guid id, [FromHeader] JwtSecurityToken accessToken)
         {
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+                return NotFound();
+            
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
+            
+            var photoContent = await photosRepository.GetPhotoContentAsync(id);
+            if (photoContent == null)
+                return NotFound();
+
+            return File(photoContent.Content, photoContent.ContentType, photoContent.FileName);
+        }
+        
+        [AllowAnonymous]
+        [HttpGet("{id}/signed-content")]
+        public async Task<IActionResult> GetPhotoSignedContent(Guid id)
+        {
+            var currentUrl = HttpContext.Request.GetEncodedUrl();
+            var check = SignedUrlHelpers.CheckSignedUrl(currentUrl);
+            if (!check)
+                return Forbid();
+    
+            var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
+            if (photoEntity == null)
+                return NotFound();
+
             var photoContent = await photosRepository.GetPhotoContentAsync(id);
             if (photoContent == null)
                 return NotFound();
@@ -55,8 +94,11 @@ namespace PhotosService.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo)
+        public async Task<IActionResult> AddPhoto(PhotoToAddDto photo, [FromHeader] JwtSecurityToken accessToken)
         {
+            if (accessToken.Subject != photo.OwnerId)
+                return Forbid();
+            
             var content = Convert.FromBase64String(photo.Base64Content);
             var result = await photosRepository.AddPhotoAsync(photo.Title, photo.OwnerId, content);
             if (!result)
@@ -65,11 +107,14 @@ namespace PhotosService.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo)
+        public async Task<IActionResult> UpdatePhoto(Guid id, PhotoToUpdateDto photo, [FromHeader] JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
+            
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
 
             photoEntity.Title = photo.Title;
             var result = await photosRepository.UpdatePhotoAsync(photoEntity);
@@ -79,11 +124,14 @@ namespace PhotosService.Controllers
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeletePhoto(Guid id)
+        public async Task<IActionResult> DeletePhoto(Guid id, [FromHeader] JwtSecurityToken accessToken)
         {
             var photoEntity = await photosRepository.GetPhotoMetaAsync(id);
             if (photoEntity == null)
                 return NotFound();
+            
+            if (accessToken.Subject != photoEntity.OwnerId)
+                return Forbid();
 
             var result = await photosRepository.DeletePhotoAsync(photoEntity);
             if (!result)
@@ -93,11 +141,14 @@ namespace PhotosService.Controllers
 
         private string GeneratePhotoUrl(PhotoDto photo)
         {
-            var relativeUrl = Url.Action(nameof(GetPhotoContent), new {
+            var relativeUrl = Url.Action(nameof(GetPhotoSignedContent), new {
                 id = photo.Id
             });
             var url = "https://localhost:6001" + relativeUrl;
-            return url;
+
+            var nowUtc = DateTime.UtcNow;
+            var signedUrl = SignedUrlHelpers.CreateSignedUrl(url, nowUtc, nowUtc.AddMinutes(5));
+            return signedUrl;
         }
     }
 }
